Move NameTag into shared components dir
