+++ src/applications/personalization/components/NameTag.jsx
@@ -3,10 +3,10 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import orderBy from 'lodash/orderBy';
 
-import prefixUtilityClasses from 'platform/utilities/prefix-utility-classes';
+import prefixUtilityClasses from '~/platform/utilities/prefix-utility-classes';
 
-import { SERVICE_BADGE_IMAGE_PATHS } from '../constants';
-import { getServiceBranchDisplayName } from '../helpers';
+import { SERVICE_BADGE_IMAGE_PATHS } from '../profile/constants';
+import { getServiceBranchDisplayName } from '../profile/helpers';
 
 const NameTag = ({
   userFullName: { first, middle, last, suffix },
